% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{Test function}
\usage{
hello2(x, y)
}
\arguments{
\item{x}{parameterdescription x}

\item{y}{parameterdescription y}
}
\value{
nothing
}
\description{
test function description
}
\examples{
hello2()
hello2(1)

}
