% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_document_package.R
\name{pretty_document_package}
\alias{pretty_document_package}
\title{Pretty document a package}
\usage{
pretty_document_package(package.path,
  output.dir = file.path("pretty_man"), verbose = TRUE,
  use.existing.rds = FALSE)
}
\arguments{
\item{package.path}{Where is the package located?}

\item{output.dir}{Where should the pretty manpages be saved? Usually in "package/inst/pretty_man"

file.path("inst", "pretty_man") has gone deprecated because the package size goes
then quite up.}

\item{verbose}{Should I give my current status?}
}
\value{
invisible(all_rd_files) # All rd-files which were completed
}
\description{
Document scripts-manpages in .html \emph{with} running the examples and giving the
output for a complete package.
}
\examples{
# Somehow I could improve here using this dontrun thingy
# https://cran.r-project.org/web/packages/roxygen2/vignettes/rd.html
#  search "dontrun"
# pretty_document_package(package.path = ".")
}
