% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_document_script.R
\name{pretty_document_script}
\alias{pretty_document_script}
\title{Pretty document scripts}
\usage{
pretty_document_script(path.R.script, outputfile, currentpackage,
  check_package = FALSE)
}
\arguments{
\item{path.R.script}{Path to the R-script}

\item{outputfile}{Where should the output html be saved?}

\item{currentpackage}{If currentpackage is supplied, currentpackage is loaded before every example.

Maybe not necessary}

\item{check_package}{Should the function be checked? Default false because it takes much time and
I assume that the package has to be built in other ways.}
}
\value{
invisible(NULL)
}
\description{
Document scripts in .html \emph{with} running the examples and giving the output
}
\details{
I understand that this is no wanted standard behaviour because this takes much
space.
}
\examples{

pretty_document_script(
	path.R.script = file.path("PrettyManpage", "hello2.R") #"R/hello2.R"
	,outputfile = "prettyhelp/hello2.html"
	,check_package = FALSE)

}
